+++ web/vite.config.js
@@ -4,7 +4,7 @@ import { svelte } from '@sveltejs/vite-plugin-svelte';
 export default defineConfig({
   plugins: [svelte()],
   resolve: {
-    conditions: ['browser'],
+    conditions: ['module', 'browser', 'production'],
   },
   test: {
     environment: 'jsdom',